+++ darts/models/forecasting/catboost_model.py
@@ -132,8 +132,7 @@ class CatBoostModel(SKLearnModelWithCategoricalCovariates):
         quantiles
             Fit the model to these quantiles if the `likelihood` is set to `quantile`.
         random_state
-            Control the randomness in the fitting procedure and for sampling.
-            Default: ``None``.
+            Controls the randomness for reproducible forecasting.
         multi_models
             If True, a separate model will be trained for each future lag to predict. If False, a single model
             is trained to predict all the steps in 'output_chunk_length' (features lags are shifted back by
@@ -216,7 +215,6 @@ class CatBoostModel(SKLearnModelWithCategoricalCovariates):
         self._likelihood = _get_likelihood(
             likelihood=likelihood,
             n_outputs=output_chunk_length if multi_models else 1,
-            random_state=random_state,
             quantiles=quantiles,
         )
 
@@ -237,6 +235,7 @@ class CatBoostModel(SKLearnModelWithCategoricalCovariates):
             categorical_past_covariates=categorical_past_covariates,
             categorical_future_covariates=categorical_future_covariates,
             categorical_static_covariates=categorical_static_covariates,
+            random_state=random_state,
         )
 
         # if no loss provided, get the default loss from the model