+++ darts/models/forecasting/kalman_forecaster.py
@@ -21,16 +21,19 @@ from darts.models.forecasting.forecasting_model import (
     TransferableFutureCovariatesLocalForecastingModel,
 )
 from darts.timeseries import TimeSeries
+from darts.utils.utils import random_method
 
 logger = get_logger(__name__)
 
 
 class KalmanForecaster(TransferableFutureCovariatesLocalForecastingModel):
+    @random_method
     def __init__(
         self,
         dim_x: int = 1,
         kf: Optional[Kalman] = None,
         add_encoders: Optional[dict] = None,
+        random_state: Optional[int] = None,
     ):
         """Kalman filter Forecaster
 
@@ -79,6 +82,8 @@ class KalmanForecaster(TransferableFutureCovariatesLocalForecastingModel):
                     'tz': 'CET'
                 }
             ..
+        random_state
+            Controls the randomness for reproducible forecasting.
 
         Examples
         --------
@@ -126,13 +131,22 @@ class KalmanForecaster(TransferableFutureCovariatesLocalForecastingModel):
         predict_likelihood_parameters: bool = False,
         verbose: bool = False,
         show_warnings: bool = True,
+        random_state: Optional[int] = None,
         **kwargs,
     ) -> TimeSeries:
         # we override `predict()` to pass a non-None `series`, so that historic_future_covariates
         # will be passed to `_predict()`
         series = series if series is not None else self.training_series
-        return super().predict(n, series, future_covariates, num_samples, **kwargs)
+        return super().predict(
+            n,
+            series,
+            future_covariates,
+            num_samples,
+            random_state=random_state,
+            **kwargs,
+        )
 
+    @random_method
     def _predict(
         self,
         n: int,
@@ -142,6 +156,7 @@ class KalmanForecaster(TransferableFutureCovariatesLocalForecastingModel):
         num_samples: int = 1,
         predict_likelihood_parameters: bool = False,
         verbose: bool = False,
+        random_state: Optional[int] = None,
     ) -> TimeSeries:
         super()._predict(
             n, series, historic_future_covariates, future_covariates, num_samples