+++ darts/models/forecasting/sf_model.py
@@ -101,7 +101,7 @@ class StatsForecastModel(TransferableFutureCovariatesLocalForecastingModel):
             Optionally, produce quantile predictions at `quantiles` levels when performing probabilistic forecasting
             with `num_samples > 1` or `predict_likelihood_parameters=True`.
         random_state
-            Control the randomness of probabilistic conformal forecasts (sample generation) across different runs.
+            Controls the randomness for reproducible forecasting.
 
         Examples
         --------
@@ -169,6 +169,7 @@ class StatsForecastModel(TransferableFutureCovariatesLocalForecastingModel):
         num_samples: int = 1,
         predict_likelihood_parameters: bool = False,
         verbose: bool = False,
+        random_state: Optional[int] = None,
     ) -> TimeSeries:
         super()._predict(
             n, series, historic_future_covariates, future_covariates, num_samples
@@ -200,6 +201,7 @@ class StatsForecastModel(TransferableFutureCovariatesLocalForecastingModel):
                     n=n,
                     num_samples=num_samples,
                     predict_likelihood_parameters=predict_likelihood_parameters,
+                    random_state=random_state,
                     verbose=verbose,
                 )
             )
@@ -210,6 +212,11 @@ class StatsForecastModel(TransferableFutureCovariatesLocalForecastingModel):
             if num_samples > 1 or predict_likelihood_parameters
             else None
         )
+
+        # a seed is manually set in AutoCES and AutoTheta forecast functions which impacts the current random state, so
+        # here we save the current random state to restore if after getting quantiles
+        random_state_ = np.random.get_state()
+
         model_output = self._estimator_predict(
             n=n,
             series=series,
@@ -217,6 +224,9 @@ class StatsForecastModel(TransferableFutureCovariatesLocalForecastingModel):
             future_covariates=future_covariates,
             levels=levels,
         )
+
+        # random state restored
+        np.random.set_state(random_state_)
 
         series = series if series is not None else self.training_series