+++ darts/models/forecasting/xgboost.py
@@ -146,8 +146,7 @@ class XGBModel(SKLearnModel):
         quantiles
             Fit the model to these quantiles if the `likelihood` is set to `quantile`.
         random_state
-            Control the randomness in the fitting procedure and for sampling.
-            Default: ``None``.
+            Controls the randomness for reproducible forecasting.
         multi_models
             If True, a separate model will be trained for each future lag to predict. If False, a single model
             is trained to predict all the steps in 'output_chunk_length' (features lags are shifted back by
@@ -206,7 +205,6 @@ class XGBModel(SKLearnModel):
         self._likelihood = _get_likelihood(
             likelihood=likelihood,
             n_outputs=output_chunk_length if multi_models else 1,
-            random_state=random_state,
             quantiles=quantiles,
         )
 
@@ -220,6 +218,7 @@ class XGBModel(SKLearnModel):
             multi_models=multi_models,
             model=xgb.XGBRegressor(**self.kwargs),
             use_static_covariates=use_static_covariates,
+            random_state=random_state,
         )
 
     def fit(